changed BitInt to BigDecimal

+++ bybit-client/src/main/kotlin/io/xooxo/bybit/model/Trade.kt
@@ -1,6 +1,6 @@
 package io.xooxo.bybit.model
 
-import java.math.BigInteger
+import java.math.BigDecimal
 import java.util.*
 
 data class TradeMessage(val topic: String, val data: Array<Trade>)
@@ -10,7 +10,7 @@ data class Trade(
     val symbol: String,
     val side: String,
     val size: Long,
-    val price: BigInteger,
+    val price: BigDecimal,
     val tick_direction: String,
     val trade_id: String,
     val cross_seq: Long
